backend/requirements.txt
fastapi==0.109.2  # Рабочая версия для Python 3.11
uvicorn==0.27.0
openai-whisper==20240930
ffmpeg-python==0.2.0
# torch==2.2.1+cpu --index-url https://download.pytorch.org/whl/cpu
# torchaudio==2.2.1+cpu --index-url https://download.pytorch.org/whl/cpu
# numpy==1.24.4
langchain==0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
typing-extensions>=4.7.0
pytest==7.4.3
pytest-asyncio==0.21.0
ollama==0.4.8
